--- src/dtos/mats-method-update.dto.ts
@@ -1,3 +1,0 @@
-import { MatsMethodBaseDTO } from './mats-method-base.dto';
-
-export class UpdateMatsMethodDTO extends MatsMethodBaseDTO {}
