--- src/dtos/monitor-default-update.dto.ts
@@ -1,3 +1,0 @@
-import { MonitorDefaultBaseDTO } from './monitor-default-base.dto';
-
-export class UpdateMonitorDefaultDTO extends MonitorDefaultBaseDTO {}
