--- src/dtos/monitor-formula-update.dto.ts
@@ -1,3 +1,0 @@
-import { MonitorFormulaBaseDTO } from './monitor-formula-base.dto';
-
-export class UpdateMonitorFormulaDTO extends MonitorFormulaBaseDTO {}
