--- src/dtos/monitor-span-update.dto.ts
@@ -1,3 +1,0 @@
-import { MonitorSpanBaseDTO } from './monitor-span-base.dto';
-
-export class UpdateMonitorSpanDTO extends MonitorSpanBaseDTO {}
