--- src/dtos/unit-fuel-update.dto.ts
@@ -1,3 +1,0 @@
-import { UnitFuelBaseDTO } from './unit-fuel-base.dto';
-
-export class UpdateUnitFuelDTO extends UnitFuelBaseDTO {}
