+++ src/mats-method-workspace/mats-method.service.ts
@@ -6,7 +6,7 @@ import { MatsMethodWorkspaceRepository } from './mats-method.repository';
 import { MatsMethodMap } from '../maps/mats-method.map';
 import { MatsMethodDTO } from '../dtos/mats-method.dto';
 import { MatsMethod } from '../entities/workspace/mats-method.entity';
-import { UpdateMatsMethodDTO } from '../dtos/mats-method-update.dto';
+import { MatsMethodBaseDTO } from '../dtos/mats-method.dto';
 import { Logger } from '@us-epa-camd/easey-common/logger';
 import { MonitorPlanWorkspaceService } from '../monitor-plan-workspace/monitor-plan.service';
 
@@ -39,7 +39,7 @@ export class MatsMethodWorkspaceService {
 
   async createMethod(
     locationId: string,
-    payload: UpdateMatsMethodDTO,
+    payload: MatsMethodBaseDTO,
     userId: string,
   ): Promise<MatsMethodDTO> {
     const method = this.repository.create({
@@ -65,7 +65,7 @@ export class MatsMethodWorkspaceService {
   async updateMethod(
     methodId: string,
     locationId: string,
-    payload: UpdateMatsMethodDTO,
+    payload: MatsMethodBaseDTO,
     userId: string,
   ): Promise<MatsMethodDTO> {
     const method = await this.getMethod(methodId);
