+++ src/monitor-default-workspace/monitor-default.service.ts
@@ -6,7 +6,7 @@ import { MonitorDefaultWorkspaceRepository } from './monitor-default.repository'
 import { MonitorDefaultDTO } from '../dtos/monitor-default.dto';
 import { MonitorDefaultMap } from '../maps/monitor-default.map';
 import { MonitorDefault } from '../entities/workspace/monitor-default.entity';
-import { UpdateMonitorDefaultDTO } from '../dtos/monitor-default-update.dto';
+import { MonitorDefaultBaseDTO } from '../dtos/monitor-default.dto';
 import { Logger } from '@us-epa-camd/easey-common/logger';
 import { MonitorPlanWorkspaceService } from '../monitor-plan-workspace/monitor-plan.service';
 
@@ -43,7 +43,7 @@ export class MonitorDefaultWorkspaceService {
 
   async createDefault(
     locationId: string,
-    payload: UpdateMonitorDefaultDTO,
+    payload: MonitorDefaultBaseDTO,
     userId: string,
   ): Promise<MonitorDefaultDTO> {
     const monDefault = this.repository.create({
@@ -74,7 +74,7 @@ export class MonitorDefaultWorkspaceService {
   async updateDefault(
     locationId: string,
     defaultId: string,
-    payload: UpdateMonitorDefaultDTO,
+    payload: MonitorDefaultBaseDTO,
     userId: string,
   ): Promise<MonitorDefaultDTO> {
     const monDefault = await this.getDefault(locationId, defaultId);
