+++ src/monitor-formula-workspace/monitor-formula.controller.spec.ts
@@ -3,13 +3,13 @@ import { LoggerModule } from '@us-epa-camd/easey-common/logger';
 
 import { MonitorFormulaWorkspaceService } from './monitor-formula.service';
 import { MonitorFormulaWorkspaceController } from './monitor-formula.controller';
-import { UpdateMonitorFormulaDTO } from '../dtos/monitor-formula-update.dto';
+import { MonitorFormulaBaseDTO } from '../dtos/monitor-formula-update.dto';
 import { ConfigService } from '@nestjs/config';
 import { HttpModule } from '@nestjs/axios';
 
 const locationId = 'string';
 const methodId = 'string';
-const matsFormulaPayload: UpdateMonitorFormulaDTO = {
+const matsFormulaPayload: MonitorFormulaBaseDTO = {
   formulaId: 'string',
   parameterCode: 'string',
   formulaCode: 'string',
