+++ src/monitor-formula-workspace/monitor-formula.controller.ts
@@ -16,7 +16,7 @@ import {
 
 import { MonitorFormulaDTO } from '../dtos/monitor-formula.dto';
 import { MonitorFormulaWorkspaceService } from './monitor-formula.service';
-import { UpdateMonitorFormulaDTO } from '../dtos/monitor-formula-update.dto';
+import { MonitorFormulaBaseDTO } from '../dtos/monitor-formula.dto';
 import { AuthGuard } from '@us-epa-camd/easey-common/guards';
 import { CurrentUser } from '@us-epa-camd/easey-common/decorators/current-user.decorator';
 import { Logger } from '@us-epa-camd/easey-common/logger';
@@ -52,7 +52,7 @@ export class MonitorFormulaWorkspaceController {
   createFormula(
     @Param('locId') locationId: string,
     @Body()
-    payload: UpdateMonitorFormulaDTO,
+    payload: MonitorFormulaBaseDTO,
     @CurrentUser() userId: string,
   ): Promise<MonitorFormulaDTO> {
     this.logger.info('Creating Formula', {
@@ -74,7 +74,7 @@ export class MonitorFormulaWorkspaceController {
     @Param('locId') locationId: string,
     @Param('formulaId') formulaRecordId: string,
     @Body()
-    payload: UpdateMonitorFormulaDTO,
+    payload: MonitorFormulaBaseDTO,
     @CurrentUser() userId: string,
   ): Promise<MonitorFormulaDTO> {
     this.logger.info('Updating Formula', {
