+++ src/unit-control-workspace/unit-control.controller.ts
@@ -13,7 +13,7 @@ import {
   ApiBearerAuth,
   ApiSecurity,
 } from '@nestjs/swagger';
-import { UpdateUnitControlDTO } from '../dtos/unit-control-update.dto';
+import { UnitControlBaseDTO } from '../dtos/unit-control.dto';
 import { UnitControlDTO } from '../dtos/unit-control.dto';
 
 import { UnitControlWorkspaceService } from './unit-control.service';
@@ -56,7 +56,7 @@ export class UnitControlWorkspaceController {
     @Param('locId') locId: string,
     @Param('unitId') unitId: number,
     @Param('unitControlId') unitControlId: string,
-    @Body() payload: UpdateUnitControlDTO,
+    @Body() payload: UnitControlBaseDTO,
   ): Promise<UnitControlDTO> {
     this.Logger.info('Updating Unit Control', {
       userId,
@@ -85,7 +85,7 @@ export class UnitControlWorkspaceController {
     @CurrentUser() userId: string,
     @Param('locId') locId: string,
     @Param('unitId') unitId: number,
-    @Body() payload: UpdateUnitControlDTO,
+    @Body() payload: UnitControlBaseDTO,
   ): Promise<UnitControlDTO> {
     this.Logger.info('Creating Unit Control', {
       userId,
