+++ src/unit-control-workspace/unit-control.service.spec.ts
@@ -5,7 +5,7 @@ import { LoggerModule } from '@us-epa-camd/easey-common/logger';
 import { UnitControlMap } from '../maps/unit-control.map';
 import { UnitControlWorkspaceService } from './unit-control.service';
 import { UnitControlWorkspaceRepository } from './unit-control.repository';
-import { UpdateUnitControlDTO } from '../dtos/unit-control-update.dto';
+import { UnitControlBaseDTO } from '../dtos/unit-control-update.dto';
 import { UnitControl } from '../entities/workspace/unit-control.entity';
 import { UnitControlDTO } from '../dtos/unit-control.dto';
 import { MonitorPlanWorkspaceService } from '../monitor-plan-workspace/monitor-plan.service';
@@ -20,7 +20,7 @@ const userId = 'testuser';
 const returnedUnitControls: UnitControlDTO[] = [];
 const returnedUnitControl: UnitControlDTO = new UnitControlDTO();
 
-const payload: UpdateUnitControlDTO = {
+const payload: UnitControlBaseDTO = {
   controlCode: 'PAX',
   controlEquipParamCode: 'DL',
   installDate: new Date(Date.now()),
